Use 'controller as' syntax in the template to define seperate instances of the controller. Use <ng-controller> in template to define a new scope for each controller.
In tandem this allows for components using the same directive & controller to be configured in different ways.

Was not able to figure out how to "properly" refresh a component after a change in config...

Working currently by using <ng-if> with a scope variable 'rendered'.  When a refresh is desired, rendered is set false (causing the component to be removed).  A function is scheduled to run 1 second later that sets rendered = true and calls $scope.$apply and the component is reloaded with the change in config.