Interesting behavior for injecting services through angular.extend($scope, serviceObject)
The properties of the service which are injected are not passed by reference.  When the controller's scope is extended by the service object, a reference to the private internal variable _timestamp is added to the scope as $scope.timestamp (along with the setter/getter functions designed allow access to the private variable).

However, calling $scope.setTimestamp(timestamp) results in $scope.timestamp = undefined (seen in chrome batarang, angularJS properties in the elements tab).  Despite this, calling $scope.getTimestamp() does NOT return undefined - instead it returns the Date object which was the argument to setTimestamp.

This implies that the private variable _timestamp was injected into the controller but is not referencing the factory service object (timestamp).