Further indication that injection with angular.extend($scope, serviceObject) does not leave references to the factory object.
In this case, the service object initializes _timestamp (and timestamp, by reference) to the javascript epoch.

After calling $scope.setTimestamp(new Date()) and following with console.log($scope.getTimestamp()) it can be seen that the setter/getter is correctly storing the current date into a private variable which we cannot see in the angular js properties view in chrome.  The references passed by the service object ($scope._timestamp and $scope.timestamp remain the javascript epoch, indicating that the setter had no effect on these variables and references to them are effectively removed.)