Previous conclusion that functionality was being inherited without reference to the original factory object was incorrect as the testing was flawed...
Current implementation seems to have resolved the issue. Previously, when clicking change1 and change2 back to back the console log would be :

before:
javascript epoch
after:
t1
before:
t1
after:
t2

THe fact that following the second 'before' when switching to change2 is NOT the javascript epoch, it was realized that execution of change1 was in fact resulting in a change to the data stored in the second controller.

In the current implementation, log when clicking change1 followed by change2 is now properly:

before:
javascript epoch
after:
t1
before:
javascript epoch
after:
t2