fixed the controller name spacing... now directive template properly refers to heat1.variable instead of just variable (indicating that variable should be on the scope).
heat1 is namespaced for the controller which is defined to be 'controller-as = heat1' and anything declared 'this.variable' in the controller can be referenced in the angular html markup.

Note - in the controller the 'this' object is captured in the variable 'vm' because 'this' may have different context in execution.