modified data provider for heatmap to use a factory -> service -> controller paterrn...
factory defines common functions which will exist as singletons.  service defines instance variables which will be new'd each time the service is injected to the controller and extends itself with the factory object.

controller extends itself with the service that has been injected.

the result is the controller will have 'new instance' variables which are defined in the service, and shared singleton functions/variables defined in the factory.